SVG 图片打开OK

+++ youranxue-domain/src/main/java/com/youranxue/domain/vo/Chapter.java
@@ -11,7 +11,7 @@ import lombok.Data;
 
 @Data
 public class Chapter {
-	private long chapterId;
+	private Integer chapterId;
 	private String chapterName;
 	private List<Section> sectionList = new ArrayList<>();
